feat(console): ship production invocation UI
