+++ packages/vite-hub/src/console/runtime/client/app.vue
@@ -1,0 +1,7 @@
+<template>
+  <RouterView />
+</template>
+
+<script setup lang="ts">
+import { RouterView } from "vue-router";
+</script>
