+++ packages/vite-hub/src/console/runtime/client/styles.css
@@ -1,0 +1,47 @@
+@import "tailwindcss";
+@import "@nuxt/ui";
+
+:root {
+  background: var(--ui-bg);
+  color: var(--ui-text);
+  font-family: Inter, ui-sans-serif, system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI", sans-serif;
+}
+
+* {
+  box-sizing: border-box;
+}
+
+html,
+body,
+#app {
+  height: 100%;
+  margin: 0;
+  min-width: 320px;
+  overflow: hidden;
+}
+
+body {
+  background: var(--ui-bg);
+  color: var(--ui-text);
+  min-height: 100vh;
+}
+
+button,
+a {
+  -webkit-tap-highlight-color: transparent;
+}
+
+.vitehub-console {
+  height: 100dvh;
+  min-height: 32rem;
+  overflow: hidden;
+}
+
+.vitehub-console [data-slot="invocation"],
+.vitehub-console [data-slot="invocation-inspector"] {
+  height: 100%;
+}
+
+.vitehub-console [data-slot="invocation-inspector"] {
+  border-inline-start: 0;
+}
