+++ packages/vite-hub/src/console/vite.ts
@@ -1,13 +1,91 @@
+import { mkdir, readFile, writeFile } from "node:fs/promises"
+import { join, resolve } from "node:path"
+import { fileURLToPath } from "node:url"
+
+import { resolveViteHubProjectRoot } from "@vite-hub/internal/build/vite"
+
 import type { Plugin } from "vite"
 
 const frameworkAgentSpecifier = "vite-hub/agent"
+const consoleRuntimeRoot = fileURLToPath(new URL("./runtime", import.meta.url))
+const consolePublicRoot = join(consoleRuntimeRoot, "public/console")
+const generatedConsolePlugin = ".vitehub/nitro/console/plugin.mjs"
+
+type ConsoleNitroConfig = {
+  handlers?: Array<{ handler: string, route: string }>
+  plugins?: string[]
+  publicAssets?: Array<{ baseURL?: string, dir: string, fallthrough?: boolean }>
+  [key: string]: unknown
+}
+
+function renderConsoleNitroPlugin(projectRoot: string): string {
+  return [
+    `import { installConsoleInvocations } from "vite-hub/console/server"`,
+    `installConsoleInvocations(${JSON.stringify(projectRoot)})`,
+    "export default function viteHubConsolePlugin() {}",
+    "",
+  ].join("\n")
+}
+
+async function writeConsoleNitroPlugin(file: string, projectRoot: string): Promise<void> {
+  const contents = renderConsoleNitroPlugin(projectRoot)
+  if (await readFile(file, "utf8").catch(() => undefined) === contents) return
+  await mkdir(resolve(file, ".."), { recursive: true })
+  await writeFile(file, contents, "utf8")
+}
+
+function generatedRegistration(value: unknown, path: string): boolean {
+  return typeof value === "string"
+    && (value === path || value.replaceAll("\\", "/").endsWith(`/${path}`))
+}
+
+export function consoleVitePlugin(): Plugin {
+  return {
+    name: "vite-hub/console",
+    async config(config) {
+      const root = resolve(config.root || process.cwd())
+      const projectRoot = resolveViteHubProjectRoot(root)
+      const plugin = resolve(root, generatedConsolePlugin)
+      await writeConsoleNitroPlugin(plugin, projectRoot)
+
+      const viteConfig = config as typeof config & { nitro?: ConsoleNitroConfig }
+      const nitro: ConsoleNitroConfig = viteConfig.nitro && typeof viteConfig.nitro === "object"
+        ? { ...viteConfig.nitro }
+        : {}
+      const handlers = Array.isArray(nitro.handlers)
+        ? nitro.handlers.filter(handler => ![
+            join(consoleRuntimeRoot, "server/invocation.get.js"),
+            join(consoleRuntimeRoot, "server/invocations.get.js"),
+            join(consoleRuntimeRoot, "server/page.get.js"),
+          ].includes(handler?.handler))
+        : []
+      handlers.push(
+        { handler: join(consoleRuntimeRoot, "server/invocations.get.js"), route: "/api/_vitehub/console/invocations" },
+        { handler: join(consoleRuntimeRoot, "server/invocation.get.js"), route: "/api/_vitehub/console/invocations/:id" },
+        { handler: join(consoleRuntimeRoot, "server/page.get.js"), route: "/_vitehub" },
+        { handler: join(consoleRuntimeRoot, "server/page.get.js"), route: "/_vitehub/**" },
+      )
+      const plugins = Array.isArray(nitro.plugins)
+        ? nitro.plugins.filter(candidate => !generatedRegistration(candidate, generatedConsolePlugin))
+        : []
+      plugins.push(plugin)
+      const publicAssets = Array.isArray(nitro.publicAssets)
+        ? nitro.publicAssets.filter(asset => asset?.baseURL !== "/_vitehub/assets")
+        : []
+      publicAssets.push({ baseURL: "/_vitehub/assets", dir: consolePublicRoot, fallthrough: false })
+
+      viteConfig.nitro = { ...nitro, handlers, plugins, publicAssets }
+    },
+  }
+}
 
 function normalizeModuleId(id: string): string {
   return id.replace(/\\/g, "/").split("?", 1)[0]!
 }
 
-export function consoleInvocationRootPlugin(projectRoot: string): Plugin {
+export function consoleInvocationRootPlugin(configuredProjectRoot?: string): Plugin {
   const frameworkAgentEntries = new Set<string>()
+  let projectRoot = configuredProjectRoot
 
   function rememberFrameworkAgent(id: string): void {
     frameworkAgentEntries.add(normalizeModuleId(id))
@@ -20,6 +98,9 @@ export function consoleInvocationRootPlugin(projectRoot: string): Plugin {
     configEnvironment(_name, config) {
       if (config.consumer !== "server") return
       return { resolve: { noExternal: ["vite-hub"] } }
+    },
+    configResolved(config) {
+      projectRoot ||= resolveViteHubProjectRoot(config.root)
     },
     async buildStart() {
       const resolved = await this.resolve(frameworkAgentSpecifier, undefined, { skipSelf: true })
@@ -35,6 +116,7 @@ export function consoleInvocationRootPlugin(projectRoot: string): Plugin {
     },
     transform(code, id) {
       if (!frameworkAgentEntries.has(normalizeModuleId(id))) return
+      if (!projectRoot) this.error("[vitehub] Could not resolve the project root for the Agent invocation console.")
       return `globalThis[Symbol.for("vitehub.console.invocations.root")] = ${JSON.stringify(projectRoot)}\n${code}`
     },
   }
