+++ packages/vite-hub/test/package.test.ts
@@ -254,12 +254,14 @@ describe("framework package contract", () => {
     expect(readFileSync(`${packageRoot}/${manifest.bin.vitehub}`, "utf8")).toMatch(/^#!\/usr\/bin\/env node/)
     expect(readFileSync(`${packageRoot}/dist/env.d.ts`, "utf8")).toContain('import "@vite-hub/env/vite"')
     expect(readFileSync(`${packageRoot}/dist/cloudflare-types.d.ts`, "utf8")).toContain("@cloudflare/workers-types")
-    const consolePage = readFileSync(`${packageRoot}/dist/console/runtime/pages/agents.vue`, "utf8")
+    const consolePage = readFileSync(`${packageRoot}/dist/console/runtime/components/console-app.vue`, "utf8")
     expect(consolePage).toContain("AgentInvocationList")
     expect(consolePage).toContain(':retry-key="paginationRetryRevision"')
     expect(consolePage).toContain("list.loadMoreError.value")
     expect(consolePage).toContain("Retry loading older sessions")
     expect(consolePage).not.toContain("groupConsoleSessions")
+    expect(readFileSync(`${packageRoot}/dist/console/runtime/public/console/console.js`, "utf8")).toContain("ViteHub")
+    expect(readFileSync(`${packageRoot}/dist/console/runtime/public/console/console.css`, "utf8")).toContain("vitehub-console")
     expect(manifest.dependencies).toHaveProperty("@cloudflare/workers-types")
   })
